Move proposal stuff into right files.
Other cleanup such as spellos, 80 col violations, copy/pastos, get rid
of backend which is no longer meaningful etc.
