--- politeiawww/backend.go
@@ -1,791 +1,0 @@
-// Copyright (c) 2017-2019 The Decred developers
-// Use of this source code is governed by an ISC
-// license that can be found in the LICENSE file.
-
-package main
-
-import (
-	"crypto/sha256"
-	"encoding/base64"
-	"encoding/hex"
-	"encoding/json"
-	"fmt"
-	"io"
-	"net/http"
-	"strconv"
-	"strings"
-	"time"
-
-	"github.com/decred/dcrtime/merkle"
-	"github.com/decred/politeia/decredplugin"
-	pd "github.com/decred/politeia/politeiad/api/v1"
-	"github.com/decred/politeia/politeiad/api/v1/identity"
-	"github.com/decred/politeia/politeiad/cache"
-	www "github.com/decred/politeia/politeiawww/api/v1"
-	"github.com/decred/politeia/politeiawww/user"
-	"github.com/decred/politeia/util"
-)
-
-const (
-	// indexFile contains the file name of the index file
-	indexFile = "index.md"
-
-	// mdStream* indicate the metadata stream used for various types
-	mdStreamGeneral = 0 // General information for this proposal
-	mdStreamChanges = 2 // Changes to record
-	// Note that 14 is in use by the decred plugin
-	// Note that 15 is in use by the decred plugin
-
-	VersionMDStreamChanges         = 1
-	BackendProposalMetadataVersion = 1
-
-	// Route to reset password at GUI
-	ResetPasswordGuiRoute = "/password"
-)
-
-type MDStreamChanges struct {
-	Version             uint             `json:"version"`                       // Version of the struct
-	AdminPubKey         string           `json:"adminpubkey"`                   // Identity of the administrator
-	NewStatus           pd.RecordStatusT `json:"newstatus"`                     // NewStatus
-	StatusChangeMessage string           `json:"statuschangemessage,omitempty"` // Status change message
-	Timestamp           int64            `json:"timestamp"`                     // Timestamp of the change
-}
-
-type loginReplyWithError struct {
-	reply *www.LoginReply
-	err   error
-}
-
-type BackendProposalMetadata struct {
-	Version   uint64 `json:"version"`   // BackendProposalMetadata version
-	Timestamp int64  `json:"timestamp"` // Last update of proposal
-	Name      string `json:"name"`      // Generated proposal name
-	PublicKey string `json:"publickey"` // Key used for signature.
-	Signature string `json:"signature"` // Signature of merkle root
-}
-
-var (
-	// MinimumLoginWaitTime is the minimum amount of time to wait before the
-	// server sends a response to the client for the login route. This is done
-	// to prevent an attacker from executing a timing attack to determine whether
-	// the ErrorStatusInvalidEmailOrPassword response is specific to a bad email
-	// or bad password.
-	MinimumLoginWaitTime = 500 * time.Millisecond
-)
-
-// PluginSetting is a structure that holds key/value pairs of a plugin setting.
-type PluginSetting struct {
-	Key   string // Name of setting
-	Value string // Value of setting
-}
-
-// Plugin describes a plugin and its settings.
-type Plugin struct {
-	ID       string          // Identifier
-	Version  string          // Version
-	Settings []PluginSetting // Settings
-}
-
-type VoteDetails struct {
-	AuthorizeVote      www.AuthorizeVote      // Authorize vote
-	AuthorizeVoteReply www.AuthorizeVoteReply // Authorize vote reply
-	StartVote          www.StartVote          // Start vote
-	StartVoteReply     www.StartVoteReply     // Start vote reply
-}
-
-// encodeBackendProposalMetadata encodes BackendProposalMetadata into a JSON
-// byte slice.
-func encodeBackendProposalMetadata(md BackendProposalMetadata) ([]byte, error) {
-	b, err := json.Marshal(md)
-	if err != nil {
-		return nil, err
-	}
-
-	return b, nil
-}
-
-// decodeBackendProposalMetadata decodes a JSON byte slice into a
-// BackendProposalMetadata.
-func decodeBackendProposalMetadata(payload []byte) (*BackendProposalMetadata, error) {
-	var md BackendProposalMetadata
-
-	err := json.Unmarshal(payload, &md)
-	if err != nil {
-		return nil, err
-	}
-
-	return &md, nil
-}
-
-// decodeMDStreamChanges decodes a JSON byte slice into a slice of
-// MDStreamChanges.
-func decodeMDStreamChanges(payload []byte) ([]MDStreamChanges, error) {
-	var msc []MDStreamChanges
-
-	d := json.NewDecoder(strings.NewReader(string(payload)))
-	for {
-		var m MDStreamChanges
-		err := d.Decode(&m)
-		if err == io.EOF {
-			break
-		} else if err != nil {
-			return nil, err
-		}
-
-		msc = append(msc, m)
-	}
-
-	return msc, nil
-}
-
-func validateProposal(np www.NewProposal, u *user.User) error {
-	log.Tracef("validateProposal")
-
-	// Obtain signature
-	sig, err := util.ConvertSignature(np.Signature)
-	if err != nil {
-		return www.UserError{
-			ErrorCode: www.ErrorStatusInvalidSignature,
-		}
-	}
-
-	// Verify public key
-	id, err := checkPublicKey(u, np.PublicKey)
-	if err != nil {
-		return err
-	}
-
-	pk, err := identity.PublicIdentityFromBytes(id[:])
-	if err != nil {
-		return err
-	}
-
-	// Check for at least 1 markdown file with a non-empty payload.
-	if len(np.Files) == 0 || np.Files[0].Payload == "" {
-		return www.UserError{
-			ErrorCode: www.ErrorStatusProposalMissingFiles,
-		}
-	}
-
-	// verify if there are duplicate names
-	filenames := make(map[string]int, len(np.Files))
-	// Check that the file number policy is followed.
-	var (
-		numMDs, numImages, numIndexFiles      int
-		mdExceedsMaxSize, imageExceedsMaxSize bool
-		hashes                                []*[sha256.Size]byte
-	)
-	for _, v := range np.Files {
-		filenames[v.Name]++
-		var (
-			data []byte
-			err  error
-		)
-		if strings.HasPrefix(v.MIME, "image/") {
-			numImages++
-			data, err = base64.StdEncoding.DecodeString(v.Payload)
-			if err != nil {
-				return err
-			}
-			if len(data) > www.PolicyMaxImageSize {
-				imageExceedsMaxSize = true
-			}
-		} else {
-			numMDs++
-
-			if v.Name == indexFile {
-				numIndexFiles++
-			}
-
-			data, err = base64.StdEncoding.DecodeString(v.Payload)
-			if err != nil {
-				return err
-			}
-			if len(data) > www.PolicyMaxMDSize {
-				mdExceedsMaxSize = true
-			}
-		}
-
-		// Append digest to array for merkle root calculation
-		digest := util.Digest(data)
-		var d [sha256.Size]byte
-		copy(d[:], digest)
-		hashes = append(hashes, &d)
-	}
-
-	// verify duplicate file names
-	if len(np.Files) > 1 {
-		var repeated []string
-		for name, count := range filenames {
-			if count > 1 {
-				repeated = append(repeated, name)
-			}
-		}
-		if len(repeated) > 0 {
-			return www.UserError{
-				ErrorCode:    www.ErrorStatusProposalDuplicateFilenames,
-				ErrorContext: repeated,
-			}
-		}
-	}
-
-	// we expect one index file
-	if numIndexFiles == 0 {
-		return www.UserError{
-			ErrorCode:    www.ErrorStatusProposalMissingFiles,
-			ErrorContext: []string{indexFile},
-		}
-	}
-
-	if numMDs > www.PolicyMaxMDs {
-		return www.UserError{
-			ErrorCode: www.ErrorStatusMaxMDsExceededPolicy,
-		}
-	}
-
-	if numImages > www.PolicyMaxImages {
-		return www.UserError{
-			ErrorCode: www.ErrorStatusMaxImagesExceededPolicy,
-		}
-	}
-
-	if mdExceedsMaxSize {
-		return www.UserError{
-			ErrorCode: www.ErrorStatusMaxMDSizeExceededPolicy,
-		}
-	}
-
-	if imageExceedsMaxSize {
-		return www.UserError{
-			ErrorCode: www.ErrorStatusMaxImageSizeExceededPolicy,
-		}
-	}
-
-	// proposal title validation
-	name, err := getProposalName(np.Files)
-	if err != nil {
-		return err
-	}
-	if !util.IsValidProposalName(name) {
-		return www.UserError{
-			ErrorCode:    www.ErrorStatusProposalInvalidTitle,
-			ErrorContext: []string{util.CreateProposalNameRegex()},
-		}
-	}
-
-	// Note that we need validate the string representation of the merkle
-	mr := merkle.Root(hashes)
-	if !pk.VerifyMessage([]byte(hex.EncodeToString(mr[:])), sig) {
-		return www.UserError{
-			ErrorCode: www.ErrorStatusInvalidSignature,
-		}
-	}
-
-	return nil
-}
-
-// voteIsAuthorized returns whether the author of the proposal has authorized
-// an admin to start the voting period for the proposal.
-func voteIsAuthorized(avr www.AuthorizeVoteReply) bool {
-	if avr.Receipt == "" {
-		// Vote has not been authorized yet
-		return false
-	} else if avr.Action == www.AuthVoteActionRevoke {
-		// Vote authorization was revoked
-		return false
-	}
-	return true
-}
-
-// getVoteStatus returns the status for the provided vote.
-func getVoteStatus(avr www.AuthorizeVoteReply, svr www.StartVoteReply, bestBlock uint64) www.PropVoteStatusT {
-	if svr.StartBlockHeight == "" {
-		// Vote has not started. Check if it's been authorized yet.
-		if voteIsAuthorized(avr) {
-			return www.PropVoteStatusAuthorized
-		} else {
-			return www.PropVoteStatusNotAuthorized
-		}
-	}
-
-	// Vote has at least been started. Check if it has finished.
-	ee, err := strconv.ParseUint(svr.EndHeight, 10, 64)
-	if err != nil {
-		// This should not happen
-		log.Errorf("getVoteStatus: ParseUint failed on '%v': %v",
-			svr.EndHeight, err)
-		return www.PropVoteStatusInvalid
-	}
-
-	if bestBlock >= ee {
-		return www.PropVoteStatusFinished
-	}
-	return www.PropVoteStatusStarted
-}
-
-// getProposalName returns the proposal name based on the index markdown file.
-func getProposalName(files []www.File) (string, error) {
-	for _, file := range files {
-		if file.Name == indexFile {
-			return util.GetProposalName(file.Payload)
-		}
-	}
-	return "", nil
-}
-
-// convertWWWPropCreditFromDatabasePropCredit coverts a database proposal
-// credit to a v1 proposal credit.
-func convertWWWPropCreditFromDatabasePropCredit(credit user.ProposalCredit) www.ProposalCredit {
-	return www.ProposalCredit{
-		PaywallID:     credit.PaywallID,
-		Price:         credit.Price,
-		DatePurchased: credit.DatePurchased,
-		TxID:          credit.TxID,
-	}
-}
-
-// validateVoteBit ensures that bit is a valid vote bit.
-func validateVoteBit(vote www.Vote, bit uint64) error {
-	if len(vote.Options) == 0 {
-		return fmt.Errorf("vote corrupt")
-	}
-	if bit == 0 {
-		return fmt.Errorf("invalid bit 0x%x", bit)
-	}
-	if vote.Mask&bit != bit {
-		return fmt.Errorf("invalid mask 0x%x bit 0x%x",
-			vote.Mask, bit)
-	}
-
-	for _, v := range vote.Options {
-		if v.Bits == bit {
-			return nil
-		}
-	}
-
-	return fmt.Errorf("bit not found 0x%x", bit)
-}
-
-// initCommentScores populates the comment scores cache.
-func (p *politeiawww) initCommentScores() error {
-	log.Tracef("initCommentScores")
-
-	// Fetch decred plugin inventory from cache
-	ir, err := p.decredInventory()
-	if err != nil {
-		return fmt.Errorf("decredInventory: %v", err)
-	}
-
-	// XXX this could be done much more efficiently since we
-	// already have all of the like comments in the inventory
-	// repsonse, but re-using the updateCommentScore function is
-	// simplier. This only gets run on startup so I'm not that
-	// worried about performance for right now.
-	for _, v := range ir.Comments {
-		_, err := p.updateCommentScore(v.Token, v.CommentID)
-		if err != nil {
-			return fmt.Errorf("updateCommentScore: %v", err)
-		}
-	}
-
-	return nil
-}
-
-// ProcessCastVotes handles the www.Ballot call
-func (p *politeiawww) ProcessCastVotes(ballot *www.Ballot) (*www.BallotReply, error) {
-	log.Tracef("ProcessCastVotes")
-
-	challenge, err := util.Random(pd.ChallengeSize)
-	if err != nil {
-		return nil, err
-	}
-
-	payload, err := decredplugin.EncodeBallot(convertBallotFromWWW(*ballot))
-	if err != nil {
-		return nil, err
-	}
-	pc := pd.PluginCommand{
-		Challenge: hex.EncodeToString(challenge),
-		ID:        decredplugin.ID,
-		Command:   decredplugin.CmdBallot,
-		CommandID: decredplugin.CmdBallot,
-		Payload:   string(payload),
-	}
-
-	responseBody, err := p.makeRequest(http.MethodPost,
-		pd.PluginCommandRoute, pc)
-	if err != nil {
-		return nil, err
-	}
-
-	var reply pd.PluginCommandReply
-	err = json.Unmarshal(responseBody, &reply)
-	if err != nil {
-		return nil, fmt.Errorf("Could not unmarshal "+
-			"PluginCommandReply: %v", err)
-	}
-
-	// Verify the challenge.
-	err = util.VerifyChallenge(p.cfg.Identity, challenge, reply.Response)
-	if err != nil {
-		return nil, err
-	}
-
-	// Decode plugin reply
-	br, err := decredplugin.DecodeBallotReply([]byte(reply.Payload))
-	if err != nil {
-		return nil, err
-	}
-	brr := convertBallotReplyFromDecredPlugin(*br)
-	return &brr, nil
-}
-
-// ProcessAuthorizeVote sends the authorizevote command to decred plugin to
-// indicate that a proposal has been finalized and is ready to be voted on.
-func (p *politeiawww) ProcessAuthorizeVote(av www.AuthorizeVote, u *user.User) (*www.AuthorizeVoteReply, error) {
-	log.Tracef("ProcessAuthorizeVote %v", av.Token)
-
-	// Get proposal from the cache
-	pr, err := p.getProp(av.Token)
-	if err != nil {
-		if err == cache.ErrRecordNotFound {
-			err = www.UserError{
-				ErrorCode: www.ErrorStatusProposalNotFound,
-			}
-		}
-		return nil, err
-	}
-
-	// Verify signature authenticity
-	err = checkPublicKeyAndSignature(u, av.PublicKey, av.Signature,
-		av.Token, pr.Version, av.Action)
-	if err != nil {
-		return nil, err
-	}
-
-	// Get vote details from cache
-	vdr, err := p.decredVoteDetails(av.Token)
-	if err != nil {
-		return nil, fmt.Errorf("decredVoteDetails: %v", err)
-	}
-	vd := convertVoteDetailsReplyFromDecred(*vdr)
-
-	// Verify record is in the right state and that the authorize
-	// vote request is valid. A vote authorization may already
-	// exist. We also allow vote authorizations to be revoked.
-	switch {
-	case pr.Status != www.PropStatusPublic:
-		// Record not public
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusWrongStatus,
-		}
-	case vd.StartVoteReply.StartBlockHeight != "":
-		// Vote has already started
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusWrongVoteStatus,
-		}
-	case av.Action != www.AuthVoteActionAuthorize &&
-		av.Action != www.AuthVoteActionRevoke:
-		// Invalid authorize vote action
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusInvalidAuthVoteAction,
-		}
-	case av.Action == www.AuthVoteActionAuthorize &&
-		voteIsAuthorized(vd.AuthorizeVoteReply):
-		// Cannot authorize vote; vote has already been
-		// authorized
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusVoteAlreadyAuthorized,
-		}
-	case av.Action == www.AuthVoteActionRevoke &&
-		!voteIsAuthorized(vd.AuthorizeVoteReply):
-		// Cannot revoke authorization; vote has not been
-		// authorized
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusVoteNotAuthorized,
-		}
-	case pr.PublicKey != av.PublicKey:
-		// User is not the author. First make sure the author didn't
-		// submit the proposal using an old identity.
-		p.RLock()
-		userID, ok := p.userPubkeys[pr.PublicKey]
-		p.RUnlock()
-		if ok {
-			if u.ID.String() != userID {
-				return nil, www.UserError{
-					ErrorCode: www.ErrorStatusUserNotAuthor,
-				}
-			}
-		} else {
-			// This should not happen
-			return nil, fmt.Errorf("proposal author not found")
-		}
-	}
-
-	// Setup plugin command
-	challenge, err := util.Random(pd.ChallengeSize)
-	if err != nil {
-		return nil, fmt.Errorf("Random: %v", err)
-	}
-
-	dav := convertAuthorizeVoteFromWWW(av)
-	payload, err := decredplugin.EncodeAuthorizeVote(dav)
-	if err != nil {
-		return nil, fmt.Errorf("EncodeAuthorizeVote: %v", err)
-	}
-
-	pc := pd.PluginCommand{
-		Challenge: hex.EncodeToString(challenge),
-		ID:        decredplugin.ID,
-		Command:   decredplugin.CmdAuthorizeVote,
-		CommandID: decredplugin.CmdAuthorizeVote + " " + av.Token,
-		Payload:   string(payload),
-	}
-
-	// Send authorizevote plugin request
-	responseBody, err := p.makeRequest(http.MethodPost,
-		pd.PluginCommandRoute, pc)
-	if err != nil {
-		return nil, err
-	}
-
-	var reply pd.PluginCommandReply
-	err = json.Unmarshal(responseBody, &reply)
-	if err != nil {
-		return nil, fmt.Errorf("Unmarshal PluginCommandReply: %v", err)
-	}
-
-	// Verify challenge
-	err = util.VerifyChallenge(p.cfg.Identity, challenge, reply.Response)
-	if err != nil {
-		return nil, fmt.Errorf("VerifyChallenge: %v", err)
-	}
-
-	// Decode plugin reply
-	avr, err := decredplugin.DecodeAuthorizeVoteReply([]byte(reply.Payload))
-	if err != nil {
-		return nil, fmt.Errorf("DecodeAuthorizeVoteReply: %v", err)
-	}
-
-	if !p.test && avr.Action == www.AuthVoteActionAuthorize {
-		p.fireEvent(EventTypeProposalVoteAuthorized,
-			EventDataProposalVoteAuthorized{
-				AuthorizeVote: &av,
-				User:          u,
-			},
-		)
-	}
-
-	return &www.AuthorizeVoteReply{
-		Action:  avr.Action,
-		Receipt: avr.Receipt,
-	}, nil
-}
-
-// ProcessStartVote handles the www.StartVote call.
-func (p *politeiawww) ProcessStartVote(sv www.StartVote, u *user.User) (*www.StartVoteReply, error) {
-	log.Tracef("ProcessStartVote %v", sv.Vote.Token)
-
-	// Verify user
-	err := checkPublicKeyAndSignature(u, sv.PublicKey, sv.Signature,
-		sv.Vote.Token)
-	if err != nil {
-		return nil, err
-	}
-
-	// Validate vote bits
-	for _, v := range sv.Vote.Options {
-		err = validateVoteBit(sv.Vote, v.Bits)
-		if err != nil {
-			return nil, www.UserError{
-				ErrorCode: www.ErrorStatusInvalidPropVoteBits,
-			}
-		}
-	}
-
-	// Validate vote parameters
-	if sv.Vote.Duration < p.cfg.VoteDurationMin ||
-		sv.Vote.Duration > p.cfg.VoteDurationMax ||
-		sv.Vote.QuorumPercentage > 100 || sv.Vote.PassPercentage > 100 {
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusInvalidPropVoteParams,
-		}
-	}
-
-	// Create vote bits as plugin payload
-	dsv := convertStartVoteFromWWW(sv)
-	payload, err := decredplugin.EncodeStartVote(dsv)
-	if err != nil {
-		return nil, err
-	}
-
-	// Get proposal from the cache
-	pr, err := p.getProp(sv.Vote.Token)
-	if err != nil {
-		if err == cache.ErrRecordNotFound {
-			err = www.UserError{
-				ErrorCode: www.ErrorStatusProposalNotFound,
-			}
-		}
-		return nil, err
-	}
-
-	// Get vote details from cache
-	vdr, err := p.decredVoteDetails(sv.Vote.Token)
-	if err != nil {
-		return nil, fmt.Errorf("decredVoteDetails: %v", err)
-	}
-	vd := convertVoteDetailsReplyFromDecred(*vdr)
-
-	// Ensure record is public, vote has been authorized,
-	// and vote has not already started.
-	if pr.Status != www.PropStatusPublic {
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusWrongStatus,
-		}
-	}
-	if !voteIsAuthorized(vd.AuthorizeVoteReply) {
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusVoteNotAuthorized,
-		}
-	}
-	if vd.StartVoteReply.StartBlockHeight != "" {
-		return nil, www.UserError{
-			ErrorCode: www.ErrorStatusWrongVoteStatus,
-		}
-	}
-
-	// Tell decred plugin to start voting
-	challenge, err := util.Random(pd.ChallengeSize)
-	if err != nil {
-		return nil, err
-	}
-
-	pc := pd.PluginCommand{
-		Challenge: hex.EncodeToString(challenge),
-		ID:        decredplugin.ID,
-		Command:   decredplugin.CmdStartVote,
-		CommandID: decredplugin.CmdStartVote + " " + sv.Vote.Token,
-		Payload:   string(payload),
-	}
-
-	responseBody, err := p.makeRequest(http.MethodPost,
-		pd.PluginCommandRoute, pc)
-	if err != nil {
-		return nil, err
-	}
-
-	var reply pd.PluginCommandReply
-	err = json.Unmarshal(responseBody, &reply)
-	if err != nil {
-		return nil, fmt.Errorf("Could not unmarshal "+
-			"PluginCommandReply: %v", err)
-	}
-
-	// Verify the challenge.
-	err = util.VerifyChallenge(p.cfg.Identity, challenge, reply.Response)
-	if err != nil {
-		return nil, err
-	}
-
-	vr, err := decredplugin.DecodeStartVoteReply([]byte(reply.Payload))
-	if err != nil {
-		return nil, err
-	}
-
-	if !p.test {
-		p.eventManager._fireEvent(EventTypeProposalVoteStarted,
-			EventDataProposalVoteStarted{
-				AdminUser: u,
-				StartVote: &sv,
-			},
-		)
-	}
-
-	// return a copy
-	rv := convertStartVoteReplyFromDecred(*vr)
-	return &rv, nil
-}
-
-// getBestBlock asks the decred plugin what the current best block is.
-func (p *politeiawww) getBestBlock() (uint64, error) {
-	challenge, err := util.Random(pd.ChallengeSize)
-	if err != nil {
-		return 0, err
-	}
-
-	pc := pd.PluginCommand{
-		Challenge: hex.EncodeToString(challenge),
-		ID:        decredplugin.ID,
-		Command:   decredplugin.CmdBestBlock,
-		CommandID: decredplugin.CmdBestBlock,
-		Payload:   "",
-	}
-
-	responseBody, err := p.makeRequest(http.MethodPost,
-		pd.PluginCommandRoute, pc)
-	if err != nil {
-		return 0, err
-	}
-
-	var reply pd.PluginCommandReply
-	err = json.Unmarshal(responseBody, &reply)
-	if err != nil {
-		return 0, fmt.Errorf("Could not unmarshal "+
-			"PluginCommandReply: %v", err)
-	}
-
-	// Verify the challenge.
-	err = util.VerifyChallenge(p.cfg.Identity, challenge, reply.Response)
-	if err != nil {
-		return 0, err
-	}
-
-	bestBlock, err := strconv.ParseUint(reply.Payload, 10, 64)
-	if err != nil {
-		return 0, err
-	}
-
-	return bestBlock, nil
-}
-
-// getPluginInventory obtains the politeiad plugin inventory.
-func (p *politeiawww) getPluginInventory() ([]Plugin, error) {
-	log.Tracef("getPluginInventory")
-
-	// Setup politeiad request
-	challenge, err := util.Random(pd.ChallengeSize)
-	if err != nil {
-		return nil, err
-	}
-
-	pi := pd.PluginInventory{
-		Challenge: hex.EncodeToString(challenge),
-	}
-
-	// Send politeiad request
-	responseBody, err := p.makeRequest(http.MethodPost,
-		pd.PluginInventoryRoute, pi)
-	if err != nil {
-		return nil, fmt.Errorf("makeRequest: %v", err)
-	}
-
-	// Handle response
-	var reply pd.PluginInventoryReply
-	err = json.Unmarshal(responseBody, &reply)
-	if err != nil {
-		return nil, err
-	}
-
-	err = util.VerifyChallenge(p.cfg.Identity, challenge, reply.Response)
-	if err != nil {
-		return nil, err
-	}
-
-	plugins := make([]Plugin, 0, len(reply.Plugins))
-	for _, v := range reply.Plugins {
-		plugins = append(plugins, convertPluginFromPD(v))
-	}
-
-	return plugins, nil
-}
