+++ politeiawww/paywall.go
@@ -195,74 +195,6 @@ func (p *politeiawww) generateProposalPaywall(u *user.User) (*user.ProposalPaywa
 	return &pp, nil
 }
 
-// ProcessProposalPaywallDetails returns a proposal paywall that enables the
-// the user to purchase proposal credits. The user can only have one paywall
-// active at a time.  If no paywall currently exists, a new one is created and
-// the user is added to the paywall pool.
-func (p *politeiawww) ProcessProposalPaywallDetails(u *user.User) (*v1.ProposalPaywallDetailsReply, error) {
-	log.Tracef("ProcessProposalPaywallDetails")
-
-	// Ensure paywall is enabled
-	if !p.paywallIsEnabled() {
-		return &v1.ProposalPaywallDetailsReply{}, nil
-	}
-
-	// Proposal paywalls cannot be generated until the user has paid their
-	// user registration fee.
-	if !p.HasUserPaid(u) {
-		return nil, v1.UserError{
-			ErrorCode: v1.ErrorStatusUserNotPaid,
-		}
-	}
-
-	var pp *user.ProposalPaywall
-	if p.userHasValidProposalPaywall(u) {
-		// Don't create a new paywall if a valid one already exists.
-		pp = p.mostRecentProposalPaywall(u)
-	} else {
-		// Create a new paywall.
-		var err error
-		pp, err = p.generateProposalPaywall(u)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	return &v1.ProposalPaywallDetailsReply{
-		CreditPrice:        pp.CreditPrice,
-		PaywallAddress:     pp.Address,
-		PaywallTxNotBefore: pp.TxNotBefore,
-	}, nil
-}
-
-// ProcessProposalPaywallPayment checks if the user has a pending paywall
-// payment and returns the payment details if one is found.
-func (p *politeiawww) ProcessProposalPaywallPayment(u *user.User) (*v1.ProposalPaywallPaymentReply, error) {
-	log.Tracef("ProcessProposalPaywallPayment")
-
-	var (
-		txID          string
-		txAmount      uint64
-		confirmations uint64
-	)
-
-	p.RLock()
-	defer p.RUnlock()
-
-	poolMember, ok := p.userPaywallPool[u.ID]
-	if ok {
-		txID = poolMember.txID
-		txAmount = poolMember.txAmount
-		confirmations = poolMember.txConfirmations
-	}
-
-	return &v1.ProposalPaywallPaymentReply{
-		TxID:          txID,
-		TxAmount:      txAmount,
-		Confirmations: confirmations,
-	}, nil
-}
-
 // verifyPropoposalPayment checks whether a payment has been sent to the
 // user's proposal paywall address. Proposal credits are created and added to
 // the user's account if the payment meets the minimum requirements.
