+++ politeiawww/www.go
@@ -10,25 +10,20 @@ import (
 	"crypto/elliptic"
 	"crypto/tls"
 	_ "encoding/gob"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"io/ioutil"
 	"net/http"
-	"net/http/httputil"
 	"os"
 	"os/signal"
 	"path/filepath"
 	"runtime/debug"
 	"strings"
-	"sync"
 	"syscall"
 	"text/template"
 	"time"
 
-	"github.com/davecgh/go-spew/spew"
-	"github.com/decred/politeia/politeiad/api/v1/mime"
 	"github.com/decred/politeia/politeiad/cache"
 	"github.com/decred/politeia/politeiad/cache/cockroachdb"
 	v1 "github.com/decred/politeia/politeiawww/api/v1"
@@ -39,7 +34,6 @@ import (
 	"github.com/gorilla/csrf"
 	"github.com/gorilla/mux"
 	"github.com/gorilla/sessions"
-	"github.com/gorilla/websocket"
 )
 
 type permission uint
@@ -58,32 +52,6 @@ var (
 	// in a session and indicates that the user is not logged in.
 	ErrSessionUUIDNotFound = errors.New("session UUID not found")
 )
-
-// wsContext is the websocket context. If uuid == "" then it is an
-// unauthenticated websocket.
-type wsContext struct {
-	uuid          string
-	rid           string
-	conn          *websocket.Conn
-	wg            sync.WaitGroup
-	subscriptions map[string]struct{}
-	errorC        chan v1.WSError
-	pingC         chan struct{}
-	done          chan struct{} // SHUT...DOWN...EVERYTHING...
-}
-
-func (w *wsContext) String() string {
-	u := w.uuid
-	if u == "" {
-		u = "anon"
-	}
-	return u + " " + w.rid
-}
-
-// IsAuthenticated returns true if the websocket is authenticated.
-func (w *wsContext) isAuthenticated() bool {
-	return w.uuid != ""
-}
 
 // Fetch remote identity
 func (p *politeiawww) getIdentity() error {
@@ -195,906 +163,6 @@ func RespondWithError(w http.ResponseWriter, r *http.Request, userHttpCode int, 
 		v1.ErrorReply{
 			ErrorCode: errorCode,
 		})
-}
-
-// version is an HTTP GET to determine what version and API route this backend
-// is using.  Additionally it is used to obtain a CSRF token.
-func (p *politeiawww) handleVersion(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleVersion")
-
-	versionReply, err := json.Marshal(v1.VersionReply{
-		Version: v1.PoliteiaWWWAPIVersion,
-		Route:   v1.PoliteiaWWWAPIRoute,
-		PubKey:  hex.EncodeToString(p.cfg.Identity.Key[:]),
-		TestNet: p.cfg.TestNet,
-	})
-	if err != nil {
-		RespondWithError(w, r, 0, "handleVersion: Marshal %v", err)
-		return
-	}
-
-	// Check if there's an active AND invalid session.
-	session, err := p.getSession(r)
-	if err != nil && session != nil {
-		// Create and save a new session for the user.
-		session := sessions.NewSession(p.store, v1.CookieSession)
-		opts := *p.store.Options
-		session.Options = &opts
-		session.IsNew = true
-		err = session.Save(r, w)
-		if err != nil {
-			RespondWithError(w, r, 0, "handleVersion: session.Save %v", err)
-			return
-		}
-	}
-
-	w.Header().Set("Content-Type", "application/json")
-	w.Header().Add("Strict-Transport-Security",
-		"max-age=63072000; includeSubDomains")
-	w.Header().Set(v1.CsrfToken, csrf.Token(r))
-	w.WriteHeader(http.StatusOK)
-	w.Write(versionReply)
-}
-
-// handleProposalPaywallDetails returns paywall details that allows the user to
-// purchase proposal credits.
-func (p *politeiawww) handleProposalPaywallDetails(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleProposalPaywallDetails")
-
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleProposalPaywallDetails: getSessionUser %v", err)
-		return
-	}
-
-	reply, err := p.ProcessProposalPaywallDetails(user)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleProposalPaywallDetails: ProcessProposalPaywallDetails  %v", err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, reply)
-}
-
-// handleProposalPaywallPayment returns the payment details for a pending
-// proposal paywall payment.
-func (p *politeiawww) handleProposalPaywallPayment(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleProposalPaywallPayment")
-
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleProposalPaywallPayment: getSessionUser %v", err)
-		return
-	}
-
-	reply, err := p.ProcessProposalPaywallPayment(user)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleProposalPaywallPayment: ProcessProposalPaywallPayment  %v", err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, reply)
-}
-
-func (p *politeiawww) websocketPing(id string) {
-	log.Tracef("websocketPing %v", id)
-	defer log.Tracef("websocketPing exit %v", id)
-
-	p.wsMtx.RLock()
-	defer p.wsMtx.RUnlock()
-
-	for _, v := range p.ws[id] {
-		if _, ok := v.subscriptions[v1.WSCPing]; !ok {
-			continue
-		}
-
-		select {
-		case v.pingC <- struct{}{}:
-		default:
-		}
-	}
-}
-
-func (p *politeiawww) handleWebsocketRead(wc *wsContext) {
-	defer wc.wg.Done()
-
-	log.Tracef("handleWebsocketRead %v", wc)
-	defer log.Tracef("handleWebsocketRead exit %v", wc)
-
-	for {
-		cmd, id, payload, err := util.WSRead(wc.conn)
-		if err != nil {
-			log.Tracef("handleWebsocketRead read %v %v", wc, err)
-			close(wc.done) // force handlers to quit
-			return
-		}
-		switch cmd {
-		case v1.WSCSubscribe:
-			subscribe, ok := payload.(v1.WSSubscribe)
-			if !ok {
-				// We are treating this a hard error so that
-				// the client knows they sent in something
-				// wrong.
-				log.Errorf("handleWebsocketRead invalid "+
-					"subscribe type %v %v", wc,
-					spew.Sdump(payload))
-				return
-			}
-
-			//log.Tracef("subscribe: %v %v", wc.uuid,
-			//	spew.Sdump(subscribe))
-
-			subscriptions := make(map[string]struct{})
-			var errors []string
-			for _, v := range subscribe.RPCS {
-				if !util.ValidSubscription(v) {
-					log.Tracef("invalid subscription %v %v",
-						wc, v)
-					errors = append(errors,
-						fmt.Sprintf("invalid "+
-							"subscription %v", v))
-					continue
-				}
-				if util.SubsciptionReqAuth(v) &&
-					!wc.isAuthenticated() {
-					log.Tracef("requires auth %v %v", wc, v)
-					errors = append(errors,
-						fmt.Sprintf("requires "+
-							"authentication %v", v))
-					continue
-				}
-				subscriptions[v] = struct{}{}
-			}
-
-			if len(errors) == 0 {
-				// Replace old subscriptions
-				p.wsMtx.Lock()
-				wc.subscriptions = subscriptions
-				p.wsMtx.Unlock()
-			} else {
-				wc.errorC <- v1.WSError{
-					Command: v1.WSCSubscribe,
-					ID:      id,
-					Errors:  errors,
-				}
-			}
-		}
-	}
-}
-
-func (p *politeiawww) handleWebsocketWrite(wc *wsContext) {
-	defer wc.wg.Done()
-	log.Tracef("handleWebsocketWrite %v", wc)
-	defer log.Tracef("handleWebsocketWrite exit %v", wc)
-
-	for {
-		var (
-			cmd, id string
-			payload interface{}
-		)
-		select {
-		case <-wc.done:
-			return
-		case e, ok := <-wc.errorC:
-			if !ok {
-				log.Tracef("handleWebsocketWrite error not ok"+
-					" %v", wc)
-				return
-			}
-			cmd = v1.WSCError
-			id = e.ID
-			payload = e
-		case _, ok := <-wc.pingC:
-			if !ok {
-				log.Tracef("handleWebsocketWrite ping not ok"+
-					" %v", wc)
-				return
-			}
-			cmd = v1.WSCPing
-			id = ""
-			payload = v1.WSPing{Timestamp: time.Now().Unix()}
-		}
-
-		err := util.WSWrite(wc.conn, cmd, id, payload)
-		if err != nil {
-			log.Tracef("handleWebsocketWrite write %v %v", wc, err)
-			return
-		}
-	}
-}
-
-func (p *politeiawww) handleWebsocket(w http.ResponseWriter, r *http.Request, id string) {
-	log.Tracef("handleWebsocket: %v", id)
-	defer log.Tracef("handleWebsocket exit: %v", id)
-
-	// Setup context
-	wc := wsContext{
-		uuid:          id,
-		subscriptions: make(map[string]struct{}),
-		pingC:         make(chan struct{}),
-		errorC:        make(chan v1.WSError),
-		done:          make(chan struct{}),
-	}
-
-	var upgrader = websocket.Upgrader{
-		EnableCompression: true,
-	}
-
-	var err error
-	wc.conn, err = upgrader.Upgrade(w, r, nil)
-	if err != nil {
-		http.Error(w, "Could not open websocket connection",
-			http.StatusBadRequest)
-		return
-	}
-	defer wc.conn.Close() // causes read to exit as well
-
-	// Create and assign session to map
-	p.wsMtx.Lock()
-	if _, ok := p.ws[id]; !ok {
-		p.ws[id] = make(map[string]*wsContext)
-	}
-	for {
-		rid, err := util.Random(16)
-		if err != nil {
-			p.wsMtx.Unlock()
-			http.Error(w, "Could not create random session id",
-				http.StatusBadRequest)
-			return
-		}
-		wc.rid = hex.EncodeToString(rid)
-		if _, ok := p.ws[id][wc.rid]; !ok {
-			break
-		}
-	}
-	p.ws[id][wc.rid] = &wc
-	p.wsMtx.Unlock()
-
-	// Reads
-	wc.wg.Add(1)
-	go p.handleWebsocketRead(&wc)
-
-	// Writes
-	wc.wg.Add(1)
-	go p.handleWebsocketWrite(&wc)
-
-	// XXX Example of a server side notifcation. Remove once other commands
-	// can be used as examples.
-	// time.Sleep(2 * time.Second)
-	// p.websocketPing(id)
-
-	wc.wg.Wait()
-
-	// Remove session id
-	p.wsMtx.Lock()
-	delete(p.ws[id], wc.rid)
-	if len(p.ws[id]) == 0 {
-		// Remove uuid since it was the last one
-		delete(p.ws, id)
-	}
-	p.wsMtx.Unlock()
-}
-
-func (p *politeiawww) handleUnauthenticatedWebsocket(w http.ResponseWriter, r *http.Request) {
-	// We are retrieving the uuid here to make sure it is NOT set. This
-	// check looks backwards but is correct.
-	id, err := p.getSessionUUID(r)
-	if err != nil && err != ErrSessionUUIDNotFound {
-		http.Error(w, "Could not get session uuid",
-			http.StatusBadRequest)
-		return
-	}
-	if id != "" {
-		http.Error(w, "Invalid session uuid", http.StatusBadRequest)
-		return
-	}
-	log.Tracef("handleUnauthenticatedWebsocket: %v", id)
-	defer log.Tracef("handleUnauthenticatedWebsocket exit: %v", id)
-
-	p.handleWebsocket(w, r, id)
-}
-
-func (p *politeiawww) handleAuthenticatedWebsocket(w http.ResponseWriter, r *http.Request) {
-	id, err := p.getSessionUUID(r)
-	if err != nil {
-		http.Error(w, "Could not get session uuid",
-			http.StatusBadRequest)
-		return
-	}
-
-	log.Tracef("handleAuthenticatedWebsocket: %v", id)
-	defer log.Tracef("handleAuthenticatedWebsocket exit: %v", id)
-
-	p.handleWebsocket(w, r, id)
-}
-
-// handleNewProposal handles the incoming new proposal command.
-func (p *politeiawww) handleNewProposal(w http.ResponseWriter, r *http.Request) {
-	// Get the new proposal command.
-	log.Tracef("handleNewProposal")
-	var np v1.NewProposal
-	decoder := json.NewDecoder(r.Body)
-	if err := decoder.Decode(&np); err != nil {
-		RespondWithError(w, r, 0, "handleNewProposal: unmarshal", v1.UserError{
-			ErrorCode: v1.ErrorStatusInvalidInput,
-		})
-		return
-	}
-
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleNewProposal: getSessionUser %v", err)
-		return
-	}
-
-	reply, err := p.ProcessNewProposal(np, user)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleNewProposal: ProcessNewProposal %v", err)
-		return
-	}
-
-	// Reply with the challenge response and censorship token.
-	util.RespondWithJSON(w, http.StatusOK, reply)
-}
-
-// handleSetProposalStatus handles the incoming set proposal status command.
-// It's used for either publishing or censoring a proposal.
-func (p *politeiawww) handleSetProposalStatus(w http.ResponseWriter, r *http.Request) {
-	// Get the proposal status command.
-	log.Tracef("handleSetProposalStatus")
-	var sps v1.SetProposalStatus
-	decoder := json.NewDecoder(r.Body)
-	if err := decoder.Decode(&sps); err != nil {
-		RespondWithError(w, r, 0, "handleSetProposalStatus: unmarshal", v1.UserError{
-			ErrorCode: v1.ErrorStatusInvalidInput,
-		})
-		return
-	}
-
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleSetProposalStatus: getSessionUser %v", err)
-		return
-	}
-
-	// Set status
-	reply, err := p.ProcessSetProposalStatus(sps, user)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleSetProposalStatus: ProcessSetProposalStatus %v", err)
-		return
-	}
-
-	// Reply with the new proposal status.
-	util.RespondWithJSON(w, http.StatusOK, reply)
-}
-
-// handleProposalDetails handles the incoming proposal details command. It fetches
-// the complete details for an existing proposal.
-func (p *politeiawww) handleProposalDetails(w http.ResponseWriter, r *http.Request) {
-	// Add the path param to the struct.
-	log.Tracef("handleProposalDetails")
-	var pd v1.ProposalsDetails
-
-	// get version from query string parameters
-	err := util.ParseGetParams(r, &pd)
-	if err != nil {
-		RespondWithError(w, r, 0, "handleProposalDetails: ParseGetParams",
-			v1.UserError{
-				ErrorCode: v1.ErrorStatusInvalidInput,
-			})
-		return
-	}
-
-	// Get proposal token from path parameters
-	pathParams := mux.Vars(r)
-	pd.Token = pathParams["token"]
-
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		if err != ErrSessionUUIDNotFound {
-			RespondWithError(w, r, 0,
-				"handleProposalDetails: getSessionUser %v", err)
-			return
-		}
-	}
-	reply, err := p.ProcessProposalDetails(pd, user)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleProposalDetails: ProcessProposalDetails %v", err)
-		return
-	}
-
-	// Reply with the proposal details.
-	util.RespondWithJSON(w, http.StatusOK, reply)
-}
-
-func (p *politeiawww) handlePolicy(w http.ResponseWriter, r *http.Request) {
-	// Get the policy command.
-	log.Tracef("handlePolicy")
-	reply := &v1.PolicyReply{
-		MinPasswordLength:          v1.PolicyMinPasswordLength,
-		MinUsernameLength:          v1.PolicyMinUsernameLength,
-		MaxUsernameLength:          v1.PolicyMaxUsernameLength,
-		UsernameSupportedChars:     v1.PolicyUsernameSupportedChars,
-		ProposalListPageSize:       v1.ProposalListPageSize,
-		UserListPageSize:           v1.UserListPageSize,
-		MaxImages:                  v1.PolicyMaxImages,
-		MaxImageSize:               v1.PolicyMaxImageSize,
-		MaxMDs:                     v1.PolicyMaxMDs,
-		MaxMDSize:                  v1.PolicyMaxMDSize,
-		ValidMIMETypes:             mime.ValidMimeTypes(),
-		MinProposalNameLength:      v1.PolicyMinProposalNameLength,
-		MaxProposalNameLength:      v1.PolicyMaxProposalNameLength,
-		ProposalNameSupportedChars: v1.PolicyProposalNameSupportedChars,
-		MaxCommentLength:           v1.PolicyMaxCommentLength,
-	}
-	util.RespondWithJSON(w, http.StatusOK, reply)
-}
-
-// handleAllVetted replies with the list of vetted proposals.
-func (p *politeiawww) handleAllVetted(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleAllVetted")
-
-	// Get the all vetted command.
-	var v v1.GetAllVetted
-	err := util.ParseGetParams(r, &v)
-	if err != nil {
-		RespondWithError(w, r, 0, "handleAllVetted: ParseGetParams",
-			v1.UserError{
-				ErrorCode: v1.ErrorStatusInvalidInput,
-			})
-		return
-	}
-
-	vr, err := p.ProcessAllVetted(v)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleAllVetted: ProcessAllVetted %v", err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, vr)
-}
-
-// handleAllUnvetted replies with the list of unvetted proposals.
-func (p *politeiawww) handleAllUnvetted(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleAllUnvetted")
-
-	// Get the all unvetted command.
-	var u v1.GetAllUnvetted
-	err := util.ParseGetParams(r, &u)
-	if err != nil {
-		RespondWithError(w, r, 0, "handleAllUnvetted: ParseGetParams",
-			v1.UserError{
-				ErrorCode: v1.ErrorStatusInvalidInput,
-			})
-		return
-	}
-
-	ur, err := p.ProcessAllUnvetted(u)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleAllUnvetted: ProcessAllUnvetted %v", err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, ur)
-}
-
-// handleNewComment handles incomming comments.
-func (p *politeiawww) handleNewComment(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleNewComment")
-
-	var sc v1.NewComment
-	decoder := json.NewDecoder(r.Body)
-	if err := decoder.Decode(&sc); err != nil {
-		RespondWithError(w, r, 0, "handleNewComment: unmarshal",
-			v1.UserError{
-				ErrorCode: v1.ErrorStatusInvalidInput,
-			})
-		return
-	}
-
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleNewComment: getSessionUser %v", err)
-		return
-	}
-
-	cr, err := p.ProcessNewComment(sc, user)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleNewComment: ProcessNewComment: %v", err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, cr)
-}
-
-// handleLikeComment handles up or down voting of commentd.
-func (p *politeiawww) handleLikeComment(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleLikeComment")
-
-	var lc v1.LikeComment
-	decoder := json.NewDecoder(r.Body)
-	if err := decoder.Decode(&lc); err != nil {
-		RespondWithError(w, r, 0, "handleLikeComment: unmarshal",
-			v1.UserError{
-				ErrorCode: v1.ErrorStatusInvalidInput,
-			})
-		return
-	}
-
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleLikeComment: getSessionUser %v", err)
-		return
-	}
-
-	cr, err := p.ProcessLikeComment(lc, user)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleLikeComment: ProcessLikeComment %v", err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, cr)
-}
-
-// handleCensorComment handles the censoring of a comment by an admin.
-func (p *politeiawww) handleCensorComment(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleCensorComment")
-
-	var cc v1.CensorComment
-	decoder := json.NewDecoder(r.Body)
-	if err := decoder.Decode(&cc); err != nil {
-		RespondWithError(w, r, 0, "handleCensorComment: unmarshal",
-			v1.UserError{
-				ErrorCode: v1.ErrorStatusInvalidInput,
-			})
-		return
-	}
-
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleCensorComment: getSessionUser %v", err)
-		return
-	}
-
-	cr, err := p.ProcessCensorComment(cc, user)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleCensorComment: ProcessCensorComment %v", err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, cr)
-}
-
-// handleCommentsGet handles batched comments get.
-func (p *politeiawww) handleCommentsGet(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleCommentsGet")
-
-	pathParams := mux.Vars(r)
-	token := pathParams["token"]
-
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		if err != ErrSessionUUIDNotFound {
-			RespondWithError(w, r, 0,
-				"handleCommentsGet: getSessionUser %v", err)
-			return
-		}
-	}
-	gcr, err := p.ProcessCommentsGet(token, user)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleCommentsGet: ProcessCommentsGet %v", err)
-		return
-	}
-	util.RespondWithJSON(w, http.StatusOK, gcr)
-}
-
-// handleUserProposalCredits returns the spent and unspent proposal credits for
-// the logged in user.
-func (p *politeiawww) handleUserProposalCredits(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleUserProposalCredits")
-
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleUserProposalCredits: getSessionUser %v", err)
-		return
-	}
-
-	reply, err := ProcessUserProposalCredits(user)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleUserProposalCredits: ProcessUserProposalCredits  %v", err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, reply)
-}
-
-// handleUserProposals returns the proposals for the given user.
-func (p *politeiawww) handleUserProposals(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleUserProposals")
-
-	// Get the user proposals command.
-	var up v1.UserProposals
-	err := util.ParseGetParams(r, &up)
-	if err != nil {
-		RespondWithError(w, r, 0, "handleUserProposals: ParseGetParams",
-			v1.UserError{
-				ErrorCode: v1.ErrorStatusInvalidInput,
-			})
-		return
-	}
-
-	userId, err := uuid.Parse(up.UserId)
-	if err != nil {
-		RespondWithError(w, r, 0, "handleUserProposals: ParseUint",
-			v1.UserError{
-				ErrorCode: v1.ErrorStatusInvalidInput,
-			})
-		return
-	}
-
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		// since having a logged in user isn't required, simply log the error
-		log.Infof("handleUserProposals: could not get session user %v", err)
-	}
-
-	upr, err := p.ProcessUserProposals(
-		&up,
-		user != nil && user.ID == userId,
-		user != nil && user.Admin)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleUserProposals: ProcessUserProposals %v", err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, upr)
-}
-
-// handleActiveVote returns all active proposals that have an active vote.
-func (p *politeiawww) handleActiveVote(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleActiveVote")
-
-	avr, err := p.ProcessActiveVote()
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleActiveVote: ProcessActivateVote %v", err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, avr)
-}
-
-// handleCastVotes records the user votes in politeiad.
-func (p *politeiawww) handleCastVotes(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleCastVotes")
-
-	var cv v1.Ballot
-	decoder := json.NewDecoder(r.Body)
-	if err := decoder.Decode(&cv); err != nil {
-		RespondWithError(w, r, 0, "handleCastVotes: unmarshal", v1.UserError{
-			ErrorCode: v1.ErrorStatusInvalidInput,
-		})
-		return
-	}
-
-	avr, err := p.ProcessCastVotes(&cv)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleCastVotes: ProcessCastVotes %v", err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, avr)
-}
-
-// handleVoteResults returns a proposal + all voting action.
-func (p *politeiawww) handleVoteResults(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleVoteResults")
-
-	pathParams := mux.Vars(r)
-	token := pathParams["token"]
-
-	vrr, err := p.ProcessVoteResults(token)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleVoteResults: ProcessVoteResults %v",
-			err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, vrr)
-}
-
-// handleAuthorizeVote handles authorizing a proposal vote.
-func (p *politeiawww) handleAuthorizeVote(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleAuthorizeVote")
-	var av v1.AuthorizeVote
-	decoder := json.NewDecoder(r.Body)
-	if err := decoder.Decode(&av); err != nil {
-		RespondWithError(w, r, 0, "handleAuthorizeVote: unmarshal", v1.UserError{
-			ErrorCode: v1.ErrorStatusInvalidInput,
-		})
-		return
-	}
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleStartVote: getSessionUser %v", err)
-		return
-	}
-	avr, err := p.ProcessAuthorizeVote(av, user)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleStartVote: ProcessAuthorizeVote %v", err)
-		return
-	}
-	util.RespondWithJSON(w, http.StatusOK, avr)
-}
-
-// handleStartVote handles starting a vote.
-func (p *politeiawww) handleStartVote(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleStartVote")
-
-	var sv v1.StartVote
-	decoder := json.NewDecoder(r.Body)
-	if err := decoder.Decode(&sv); err != nil {
-		RespondWithError(w, r, 0, "handleStartVote: unmarshal", v1.UserError{
-			ErrorCode: v1.ErrorStatusInvalidInput,
-		})
-		return
-	}
-
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleStartVote: getSessionUser %v", err)
-		return
-	}
-
-	// Sanity
-	if !user.Admin {
-		RespondWithError(w, r, 0,
-			"handleStartVote: admin %v", user.Admin)
-		return
-	}
-
-	svr, err := p.ProcessStartVote(sv, user)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleStartVote: ProcessStartVote %v", err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, svr)
-}
-
-// handleGetAllVoteStatus returns the voting status of all public proposals.
-func (p *politeiawww) handleGetAllVoteStatus(w http.ResponseWriter, r *http.Request) {
-	gasvr, err := p.ProcessGetAllVoteStatus()
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleProposalsVotingStatus: ProcessProposalsVotingStatus %v", err)
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, gasvr)
-}
-
-// handleVoteStatus returns the vote status for a given proposal.
-func (p *politeiawww) handleVoteStatus(w http.ResponseWriter, r *http.Request) {
-	pathParams := mux.Vars(r)
-	vsr, err := p.ProcessVoteStatus(pathParams["token"])
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleCommentsGet: ProcessCommentGet %v", err)
-		return
-	}
-	util.RespondWithJSON(w, http.StatusOK, vsr)
-}
-
-// handleUserCommentsLikes returns the user votes on comments of a given proposal.
-func (p *politeiawww) handleUserCommentsLikes(w http.ResponseWriter, r *http.Request) {
-	log.Tracef("handleUserCommentsLikes")
-
-	pathParams := mux.Vars(r)
-	token := pathParams["token"]
-
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleUserCommentsLikes: getSessionUser %v", err)
-		return
-	}
-
-	uclr, err := p.ProcessUserCommentsLikes(user, token)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleUserCommentsLikes: processUserCommentsLikes %v", err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, uclr)
-}
-
-// handleEditProposal attempts to edit a proposal
-func (p *politeiawww) handleEditProposal(w http.ResponseWriter, r *http.Request) {
-	var ep v1.EditProposal
-	decoder := json.NewDecoder(r.Body)
-	if err := decoder.Decode(&ep); err != nil {
-		RespondWithError(w, r, 0, "handleEditProposal: unmarshal",
-			v1.UserError{
-				ErrorCode: v1.ErrorStatusInvalidInput,
-			})
-		return
-	}
-
-	user, err := p.getSessionUser(w, r)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleEditProposal: getSessionUser %v", err)
-		return
-	}
-
-	log.Debugf("handleEditProposal: %v", ep.Token)
-
-	epr, err := p.ProcessEditProposal(ep, user)
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleEditProposal: ProcessEditProposal %v", err)
-		return
-	}
-
-	util.RespondWithJSON(w, http.StatusOK, epr)
-}
-
-// handleProposalsStats returns the counting of proposals aggrouped by each proposal status
-func (p *politeiawww) handleProposalsStats(w http.ResponseWriter, r *http.Request) {
-	psr, err := p.ProcessProposalsStats()
-	if err != nil {
-		RespondWithError(w, r, 0,
-			"handleProposalsStats: ProcessProposalsStats %v", err)
-		return
-	}
-	util.RespondWithJSON(w, http.StatusOK, psr)
-}
-
-// handleNotFound is a generic handler for an invalid route.
-func (p *politeiawww) handleNotFound(w http.ResponseWriter, r *http.Request) {
-	// Log incoming connection
-	log.Debugf("Invalid route: %v %v %v %v", remoteAddr(r), r.Method, r.URL,
-		r.Proto)
-
-	// Trace incoming request
-	log.Tracef("%v", newLogClosure(func() string {
-		trace, err := httputil.DumpRequest(r, true)
-		if err != nil {
-			trace = []byte(fmt.Sprintf("logging: "+
-				"DumpRequest %v", err))
-		}
-		return string(trace)
-	}))
-
-	util.RespondWithJSON(w, http.StatusNotFound, v1.ErrorReply{})
 }
 
 // addRoute sets up a handler for a specific method+route. If methos is not
